data/pyproject.toml
[project]
name = "blockchain-bandits-data"
version = "0.1.0"
description = "ETL and label-source loaders"
requires-python = ">=3.11"
dependencies = [
  "google-cloud-bigquery>=3.25.0",
  "pandas>=2.2.0",
  "polars>=1.9.0",
  "pyarrow>=17.0.0",
  "duckdb>=1.1.0",
  "httpx>=0.27.0",
  "boto3>=1.35.0",
  "python-dotenv>=1.0.1",
  "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "ruff>=0.7.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bb_data"]
